go: split mempool helpers by policy area
