+++ clients/go/node/mempool_ordering.go
@@ -1,0 +1,91 @@
+package node
+
+import (
+	"bytes"
+	"math/bits"
+	"sort"
+)
+
+func (m *Mempool) SelectTransactions(maxCount int, maxBytes int) [][]byte {
+	if m == nil || maxCount <= 0 || maxBytes <= 0 {
+		return nil
+	}
+
+	entries := m.snapshotEntries()
+	sortMempoolEntries(entries)
+	return pickEntries(entries, maxCount, maxBytes)
+}
+
+func (m *Mempool) snapshotEntries() []*mempoolEntry {
+	m.mu.RLock()
+	defer m.mu.RUnlock()
+	entries := make([]*mempoolEntry, 0, len(m.txs))
+	for _, entry := range m.txs {
+		entries = append(entries, entry)
+	}
+	return entries
+}
+
+func sortMempoolEntries(entries []*mempoolEntry) {
+	sort.Slice(entries, func(i, j int) bool {
+		if cmp := compareFeeRate(entries[i], entries[j]); cmp != 0 {
+			return cmp > 0
+		}
+		if entries[i].fee != entries[j].fee {
+			return entries[i].fee > entries[j].fee
+		}
+		if entries[i].weight != entries[j].weight {
+			return entries[i].weight < entries[j].weight
+		}
+		return bytes.Compare(entries[i].txid[:], entries[j].txid[:]) < 0
+	})
+}
+
+func pickEntries(entries []*mempoolEntry, maxCount int, maxBytes int) [][]byte {
+	selected := make([][]byte, 0, len(entries))
+	usedBytes := 0
+	for _, entry := range entries {
+		if len(selected) >= maxCount {
+			break
+		}
+		if entry.size > maxBytes-usedBytes {
+			continue
+		}
+		selected = append(selected, append([]byte(nil), entry.raw...))
+		usedBytes += entry.size
+	}
+	return selected
+}
+
+func compareFeeRate(a *mempoolEntry, b *mempoolEntry) int {
+	if a == nil || b == nil {
+		return 0
+	}
+	return compareFeeRateWeightValues(a.fee, a.weight, b.fee, b.weight)
+}
+
+func compareEvictionFeeRate(a *mempoolEntry, b *mempoolEntry) int {
+	// Eviction and miner selection intentionally share the fee/weight axis.
+	return compareFeeRate(a, b)
+}
+
+func compareFeeRateWeightValues(feeA uint64, weightA uint64, feeB uint64, weightB uint64) int {
+	if weightA == 0 || weightB == 0 {
+		return 0
+	}
+	ahi, alo := bits.Mul64(feeA, weightB)
+	bhi, blo := bits.Mul64(feeB, weightA)
+	if ahi != bhi {
+		if ahi > bhi {
+			return 1
+		}
+		return -1
+	}
+	if alo != blo {
+		if alo > blo {
+			return 1
+		}
+		return -1
+	}
+	return 0
+}
